# update test times